# web/openapi.json
{
  "components": {
    "schemas": {
...
      "CommonResponse_Any_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[Any]",
        "type": "object"
      },
...
      "CommonResponse_DeleteEgressProxyResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeleteEgressProxyResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[DeleteEgressProxyResponse]",
        "type": "object"
      },
...
      "CommonResponse_EgressProxySchema_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EgressProxySchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[EgressProxySchema]",
        "type": "object"
      },
...
      "CommonResponse_QueryEgressProxiesResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueryEgressProxiesResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[QueryEgressProxiesResponse]",
        "type": "object"
      },
...
      "CommonResponse_SandboxContainerSchema_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SandboxContainerSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[SandboxContainerSchema]",
        "type": "object"
      },
...
      "CommonResponse_TestEgressProxyResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TestEgressProxyResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[TestEgressProxyResponse]",
        "type": "object"
      },
...
      "CreateEgressProxyRequest": {
        "properties": {
          "proxy_account": {
            "default": "",
            "maxLength": 255,
            "title": "Proxy Account",
            "type": "string"
          },
          "proxy_host": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Proxy Host",
            "type": "string"
          },
          "proxy_password": {
            "default": "",
            "maxLength": 512,
            "title": "Proxy Password",
            "type": "string"
          },
          "proxy_port": {
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Proxy Port",
            "type": "integer"
          },
          "proxy_type": {
            "$ref": "#/components/schemas/EgressProxyType",
            "default": "http"
          }
        },
        "required": [
          "proxy_host",
          "proxy_port"
        ],
        "title": "CreateEgressProxyRequest",
        "type": "object"
      },
...
      "DeleteEgressProxyResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id"
        ],
        "title": "DeleteEgressProxyResponse",
        "type": "object"
      },
...
      "EgressProxySchema": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "proxy_account": {
            "title": "Proxy Account",
            "type": "string"
          },
          "proxy_host": {
            "title": "Proxy Host",
            "type": "string"
          },
          "proxy_password": {
            "title": "Proxy Password",
            "type": "string"
          },
          "proxy_port": {
            "title": "Proxy Port",
            "type": "integer"
          },
          "proxy_type": {
            "$ref": "#/components/schemas/EgressProxyType"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "proxy_type",
          "proxy_host",
          "proxy_port",
          "proxy_account",
          "proxy_password",
          "created_at",
          "updated_at"
        ],
        "title": "EgressProxySchema",
        "type": "object"
      },
      "EgressProxyType": {
        "enum": [
          "http",
          "https",
          "socks5"
        ],
        "title": "EgressProxyType",
        "type": "string"
      },
...
      "QueryEgressProxiesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EgressProxySchema"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "size",
          "total",
          "items"
        ],
        "title": "QueryEgressProxiesResponse",
        "type": "object"
      },
...
      "SandboxContainerEgressMode": {
        "enum": [
          "direct",
          "proxy",
          "tor"
        ],
        "title": "SandboxContainerEgressMode",
        "type": "string"
      },
      "SandboxContainerPortMapping": {
        "properties": {
          "container_port": {
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Container Port",
            "type": "integer"
          },
          "host_port": {
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Host Port",
            "type": "integer"
          },
          "protocol": {
            "default": "tcp",
            "enum": [
              "tcp",
              "udp"
            ],
            "title": "Protocol",
            "type": "string"
          }
        },
        "required": [
          "container_port",
          "host_port"
        ],
        "title": "SandboxContainerPortMapping",
        "type": "object"
      },
      "SandboxContainerSchema": {
        "properties": {
          "container_hash": {
            "title": "Container Hash",
            "type": "string"
          },
          "container_name": {
            "title": "Container Name",
            "type": "string"
          },
          "control_proxy_host_port": {
            "title": "Control Host Port",
            "type": "integer"
          },
          "control_proxy_port": {
            "title": "Control Port",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "egress_label": {
            "title": "Egress Label",
            "type": "string"
          },
          "egress_mode": {
            "$ref": "#/components/schemas/SandboxContainerEgressMode"
          },
          "egress_proxy_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Egress Proxy Id"
          },
          "host_id": {
            "title": "Host Id",
            "type": "integer"
          },
          "host_ip_address": {
            "title": "Host Ip Address",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_id": {
            "title": "Image Id",
            "type": "integer"
          },
          "image_name": {
            "title": "Image Name",
            "type": "string"
          },
          "owner_id": {
            "title": "Owner Id",
            "type": "integer"
          },
          "owner_username": {
            "title": "Owner Username",
            "type": "string"
          },
          "port_mappings": {
            "items": {
              "$ref": "#/components/schemas/SandboxContainerPortMapping"
            },
            "title": "Port Mappings",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/SandboxContainerStatus"
          },
          "supports_tor": {
            "title": "Supports Tor",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "host_id",
          "host_ip_address",
          "container_name",
          "container_hash",
          "image_id",
          "image_name",
          "supports_tor",
          "control_proxy_port",
          "egress_mode",
          "egress_proxy_id",
          "egress_label",
          "control_proxy_host_port",
          "port_mappings",
          "status",
          "owner_id",
          "owner_username",
          "created_at",
          "updated_at"
        ],
        "title": "SandboxContainerSchema",
        "type": "object"
      },
      "SandboxContainerStatus": {
        "enum": [
          "created",
          "running",
          "stopped",
          "error"
        ],
        "title": "SandboxContainerStatus",
        "type": "string"
      },
...
      "TestEgressProxyResponse": {
        "properties": {
          "elapsed_ms": {
            "title": "Elapsed Ms",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Code"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "success",
          "elapsed_ms",
          "message"
        ],
        "title": "TestEgressProxyResponse",
        "type": "object"
      },
...
      "UpdateEgressProxyRequest": {
        "properties": {
          "proxy_account": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Account"
          },
          "proxy_host": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Host"
          },
          "proxy_password": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Password"
          },
          "proxy_port": {
            "anyOf": [
              {
                "maximum": 65535.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy Port"
          },
          "proxy_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EgressProxyType"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "UpdateEgressProxyRequest",
        "type": "object"
      },
...
      "UpdateSandboxContainerEgressRequest": {
        "properties": {
          "egress_mode": {
            "$ref": "#/components/schemas/SandboxContainerEgressMode",
            "default": "direct"
          },
          "egress_proxy_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Egress Proxy Id"
          }
        },
        "title": "UpdateSandboxContainerEgressRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AccessTokenAuth": {
        "in": "header",
        "name": "X-Z3r0-Access-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Z3r0 - A controlled multi-agent workbench for authorized security assessment, code auditing, internal review, and controlled research.",
    "version": "0.2.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/egress-proxies": {
      "get": {
        "operationId": "query_egress_proxies_route_api_egress_proxies_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "title": "Keyword",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_QueryEgressProxiesResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Query Egress Proxies Route",
        "tags": [
          "egress-proxies"
        ]
      },
      "post": {
        "operationId": "create_egress_proxy_handler_api_egress_proxies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEgressProxyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_EgressProxySchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Create Egress Proxy Handler",
        "tags": [
          "egress-proxies"
        ]
      }
    },
    "/api/egress-proxies/{id}": {
      "delete": {
        "operationId": "delete_egress_proxy_handler_api_egress_proxies__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_DeleteEgressProxyResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Egress proxy not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Delete Egress Proxy Handler",
        "tags": [
          "egress-proxies"
        ]
      },
      "patch": {
        "operationId": "update_egress_proxy_handler_api_egress_proxies__id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEgressProxyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_EgressProxySchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Egress proxy not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Update Egress Proxy Handler",
        "tags": [
          "egress-proxies"
        ]
      }
    },
    "/api/egress-proxies/{id}/test": {
      "post": {
        "operationId": "test_egress_proxy_handler_api_egress_proxies__id__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_TestEgressProxyResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Egress proxy not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Test Egress Proxy Handler",
        "tags": [
          "egress-proxies"
        ]
      }
    },
...
    "/api/sandbox-containers/{id}/egress": {
      "patch": {
        "operationId": "update_sandbox_container_egress_handler_api_sandbox_containers__id__egress_patch",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSandboxContainerEgressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_SandboxContainerSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Sandbox container not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Update Sandbox Container Egress Handler",
        "tags": [
          "sandbox-containers"
        ]
      }
    },
...
  }
}